Added fluid splitter
Redstone and basic variant
Progress on #38

+++ src/main/java/com/Da_Technomancer/crossroads/tileentities/fluid/FluidTubeTileEntity.java
@@ -85,10 +85,10 @@ public class FluidTubeTileEntity extends TileEntity implements ITickable{
 			// full, tank and pipe are not BOTH empty, capacity and contents of
 			// pipe.
 
-			long tankContent = handler.drain(Integer.MAX_VALUE, false) == null ? 0 : handler.drain(Integer.MAX_VALUE, false).amount;
-			long tankCapacity = tankContent + handler.fill(content == null ? new FluidStack(handler.drain(1, false).getFluid(), Integer.MAX_VALUE) : new FluidStack(content.getFluid(), Integer.MAX_VALUE), false);
+			long tankContent = handler.drain(Short.MAX_VALUE, false) == null ? 0 : handler.drain(Short.MAX_VALUE, false).amount;
+			long tankCapacity = tankContent + handler.fill(content == null ? new FluidStack(handler.drain(1, false).getFluid(), Short.MAX_VALUE) : new FluidStack(content.getFluid(), Short.MAX_VALUE), false);
 
-			int total = (int) Math.min((content == null ? 0 : content.amount) + tankContent, Integer.MAX_VALUE);
+			int total = (int) Math.min((content == null ? 0 : content.amount) + tankContent, Short.MAX_VALUE);
 
 			Fluid fluid = content == null ? handler.drain(1, false).getFluid() : content.getFluid();
